added routes for user
